Move Charts into folder and add prop to CandlestickChart
