+++ frontend/src/components/Charts/BarChart.vue
